default_requirements.txt
black==20.8b1
ipython==7.19.0
jedi==0.17.2
neovim==0.3.1
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
